--- templates/standalone/.gitignore
@@ -1,22 +1,0 @@
-*.iml
-*.class
-*.log
-target/
-project/target/
-project/boot/
-dist/
-boot/
-logs/
-out/
-tmp/
-projectFilesBackup/
-.history/
-.idea/
-.idea_modules/
-.DS_STORE
-.cache
-.settings
-.project
-.classpath
-version.properties
-RUNNING_PID
