--- templates/standalone/README.md
@@ -1,3 +1,0 @@
-# Play WS Standalone Client 
-
-This is a simple project to show Play WS Standalone
