--- templates/standalone/project/plugins.sbt
@@ -1,6 +1,0 @@
-logLevel := Level.Warn
-
-resolvers += Resolver.sonatypeRepo("releases")
-
-addSbtPlugin("com.gilt" % "sbt-dependency-graph-sugar" % "0.8.2")
-
