--- templates/standalone/src/main/scala/playwsclient/ScalaClient.scala
@@ -1,29 +1,0 @@
-package playwsclient
-
-import akka.actor.ActorSystem
-import akka.stream.ActorMaterializer
-import play.api.libs.ws.StandaloneWSClient
-import play.api.libs.ws.ahc.StandaloneAhcWSClient
-
-import scala.concurrent.Future
-
-object ScalaClient {
-  import scala.concurrent.ExecutionContext.Implicits._
-
-  def main(args: Array[String]): Unit = {
-    implicit val system = ActorSystem()
-    implicit val materializer = ActorMaterializer()
-    val wsClient = StandaloneAhcWSClient()
-
-    call(wsClient)
-      .andThen { case _ => wsClient.close() }
-      .andThen { case _ => system.terminate() }
-  }
-
-  def call(wsClient: StandaloneWSClient): Future[Unit] = {
-    wsClient.url("http://www.google.com").get().map { response ⇒
-      val statusText: String = response.statusText
-      println(s"Got a response $statusText")
-    }
-  }
-}
